+++ CHANGES.rst
@@ -6,6 +6,7 @@ Version: 4.0.3
 ENHANCEMENTS:
 
 * Added the 'worker_id' context.config.userdata parameter to allow users to identify which worker is executing every feature or scenario when running tests in parallel. `PR #121 <https://github.com/hrcorval/behavex/pull/121>`_
+* Added "start" and "stop" attributes to the report.json file, to indicate the start and end time of executed features, scenarios and steps.
 
 CONTRIBUTIONS: